пернес дз в папку
